requirements.txt
asgiref==3.9.1
Django==4.2.23
django-filter==25.1
django-pwa==2.0.1
djangorestframework==3.16.1
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.15.0
